feat(DTFS2-7490): refactor send email service into appropriate directory
